DDMORE-917
In order to organise nmtran file in correct way, we need to refactor
order of Theta, omega and omega blocks with respect to Etas.

Considering, we have Etas list in an order where omegas part of omega
blocks are listed before rest of omegas, we need to do following
amendments.

1. Confirm if we have Etas list in an order where omegas part of omega
blocks are listed before rest of omegas, if not organise in this order.
2. Thetas with Etas should be at top and that too in order of ordered
Etas list above. Rest of the Thetas can follow These Thetas.
3. Individual parameter definitions should be in the order of these
ordered Thetas (in turn ordered Etas).